+++ pytorch_lightning/utilities/model_utils.py
@@ -1,6 +1,31 @@
+from typing import Union, List
+
+from torch.utils.data import DataLoader
+
 from pytorch_lightning.core.lightning import LightningModule
 from pytorch_lightning.core.datamodule import LightningDataModule
-from typing import Union
+
+
+class _PatchDataLoader(object):
+    r"""
+    Callable object for patching dataloaders passed into trainer.fit().
+    Use this class to override model.*_dataloader() and be pickle-compatible.
+
+    Args:
+        dataloader: Dataloader object to return when called.
+
+    """
+
+    def __init__(self, dataloader: Union[List[DataLoader], DataLoader]):
+        self.dataloader = dataloader
+
+        # cannot pickle __code__ so cannot verify if PatchDataloader
+        # exists which shows dataloader methods have been overwritten.
+        # so, we hack it by using the string representation
+        self.patch_loader_code = str(self.__call__.__code__)
+
+    def __call__(self) -> Union[List[DataLoader], DataLoader]:
+        return self.dataloader
 
 
 def is_overridden(method_name: str, model: Union[LightningModule, LightningDataModule]) -> bool:
